backend/requirements.txt
fastapi
uvicorn[standard]
opensearch-py
pyjwt
python-multipart
apscheduler
requests